Fixed directory mode names
