+++ src/slp2mp4/modes/replay_manager.py
@@ -6,9 +6,8 @@ import zipfile
 
 import pathlib
 
-from slp2mp4.modes.mode import Mode
+from slp2mp4.modes.directory import Directory
 from slp2mp4.output import Output
-from slp2mp4.modes.mode import Mode
 import slp2mp4.util as util
 
 
@@ -19,39 +18,20 @@ def _make_tmpdir():
     return pathlib.Path(tmpdir)
 
 
-class ReplayManager(Mode):
-    def __init__(self, paths: list[pathlib.Path], output_directory: pathlib.Path):
-        self.lookups = {}
-        self.paths = self._extract_helper(paths)
-        self.output_directory = output_directory
-
-    def iterator(self, _root, path):
-        yield self.lookups[path]
-
-    def _extract_helper(self, paths):
-        for path in paths:
-            root = (
-                pathlib.Path(path.absolute().name)
-                if path.is_dir()
-                else util.get_parent_as_path(path) / path.stem
-            )
-            self._recursive_extract(root, path)
-        return self.lookups.keys()
-
-    def _recursive_extract(self, location, path, fromzip=False):
+class ReplayManager(Directory):
+    def _recursive_find(self, location, path, fromzip=False):
         if zipfile.is_zipfile(path):
             tmpdir = _make_tmpdir()
             with zipfile.ZipFile(path, "r") as zfile:
                 zfile.extractall(path=tmpdir)
-            self._recursive_extract(location, tmpdir, True)
+            self._recursive_find(location, tmpdir, True)
         else:
-            slps = list(sorted(path.glob("*.slp"), key=util.natsort))
-            if len(slps) > 0 and fromzip:
-                self.lookups[location] = (
-                    slps,
-                    location.parent,
-                    pathlib.Path(location.name),
-                )
+            self._add_slps(location, path, fromzip)
             for child in path.iterdir():
                 if child.is_dir() or zipfile.is_zipfile(child):
-                    self._recursive_extract(location / child.stem, child, fromzip)
+                    self._recursive_find(location / child.stem, child, fromzip)
+
+    def _add_slps(self, location, path, fromzip):
+        if not fromzip:
+            return
+        super()._add_slps(location, path)
